refactor(Microsoft Excel SharePoint Node): Revert Sheet/Table Get Many to independent operation files
createGetAllOperation had no precedent anywhere else in nodes-base:
duplicate "get many" operations (e.g. Excel v2's worksheet/getAll and
workbook/getAll) each keep their own properties/execute and only share
the low-level request/pagination helpers. Go back to that shape instead
of a factory that builds the whole operation.
